+++ spec/transpec/rspec_version_spec.rb
@@ -27,8 +27,10 @@ module Transpec
       include_examples 'feature availability', method, [
         ['2.14.0',       false],
         ['2.99.0.beta1', true],
+        ['2.99.0.beta2', true],
         ['2.99.0',       true],
         ['3.0.0.beta1',  true],
+        ['3.0.0.beta2',  true],
         ['3.0.0',        true]
       ]
     end
@@ -39,7 +41,8 @@ module Transpec
         ['2.99.0.beta1', false],
         ['2.99.0.beta2', true],
         ['2.99.0',       true],
-        ['3.0.0.beta1',  true],
+        ['3.0.0.beta1',  false],
+        ['3.0.0.beta2',  true],
         ['3.0.0',        true]
       ]
     end
@@ -48,8 +51,10 @@ module Transpec
       include_examples 'feature availability', method, [
         ['2.14.0',       false],
         ['2.99.0.beta1', false],
+        ['2.99.0.beta2', false],
         ['2.99.0',       false],
         ['3.0.0.beta1',  true],
+        ['3.0.0.beta2',  true],
         ['3.0.0',        true]
       ]
     end
@@ -58,6 +63,7 @@ module Transpec
       include_examples 'feature availability', method, [
         ['2.14.0',       false],
         ['2.99.0.beta1', false],
+        ['2.99.0.beta2', false],
         ['2.99.0',       false],
         ['3.0.0.beta1',  false],
         ['3.0.0.beta2',  true],